requirements.txt
# For MQTT acceptance tests
paho-mqtt
fastapi
uvicorn
pydantic
httpx
jsonschema
python-dotenv
